app-source/worker/requirements.txt
flask
redis
flask-cors
psycopg2-binary